Minimal stack inline migration
